+++ packages/component/src/block/poster/poster.stories.tsx
@@ -150,6 +150,11 @@ const meta: Meta = {
           { title: "Quantity", value: "0.123" },
         ],
       },
+      referral: {
+        code: "WRECKED",
+        link: "https://woo.org?size=100",
+        slogan: "Try WOOFi Pro now with:",
+      },
     },
     layout: {},
   },
@@ -174,7 +179,13 @@ export const Default: Story = {
     const posterRef = useRef<PosterRef | null>(null);
     return (
       <div>
-        <Poster width={500} height={300} data={args} ref={posterRef} />
+        <Poster
+          width={500}
+          height={300}
+          data={args}
+          ref={posterRef}
+          ratio={3}
+        />
         <div className="orderly-mt-3 orderly-flex orderly-gap-2">
           <Button
             size={"small"}