+++ packages/component/src/table/table.tsx
@@ -23,6 +23,7 @@ export interface TableProps<RecordType> extends TBodyProps<RecordType> {
   headerClassName?: string;
   showMaskElement?: boolean;
   loadMore?: () => void;
+  id?: string;
 }
 
 export const Table = <RecordType extends unknown>(
@@ -103,7 +104,7 @@ export const Table = <RecordType extends unknown>(
 
     if (leftDivide) {
       const left = leftDivide.getBoundingClientRect();
-      console.log("leftDivide", left);
+      // console.log("leftDivide", left);
       // setMaskLayout((layout) => ({ ...layout, left }));
     }
 
@@ -160,6 +161,7 @@ export const Table = <RecordType extends unknown>(
       canExpand={typeof props.expandRowRender === "function"}
     >
       <div
+        id={props.id}
         ref={wrapRef}
         className={cn(
           "orderly-ui-table orderly-relative orderly-flex-col orderly-overflow-x-auto orderly-peer",