+++ packages/hooks/src/referral/index.ts
@@ -5,4 +5,7 @@ export { useReferralRebateSummary } from "./useReferralRebateSummary";
 export { useRefereeHistory } from "./useRefereeHistory";
 export { useRefereeInfo } from "./useRefereeInfo";
 export { useRefereeRebateSummary } from "./useRefereeRebateSummary";
+export { useCheckReferralCode } from "./useCheckReferralCode";
+export { useGetReferralCode } from "./useGetReferralCode";
+export { useReferralInfo } from "./useReferralInfo";
 export * from "./api";