+++ packages/component/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/react
 
+## 1.5.1
+
+### Patch Changes
+
+- 34efc6c0: update
+- Updated dependencies [34efc6c0]
+  - @orderly.network/trading-view@1.5.1
+  - @orderly.network/hooks@1.5.1
+  - @orderly.network/core@1.5.1
+  - @orderly.network/net@1.5.1
+  - @orderly.network/types@1.5.1
+  - @orderly.network/utils@1.5.1
+
+## 1.5.1-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/trading-view@1.5.1-alpha.0
+  - @orderly.network/hooks@1.5.1-alpha.0
+  - @orderly.network/core@1.5.1-alpha.0
+  - @orderly.network/net@1.5.1-alpha.0
+  - @orderly.network/types@1.5.1-alpha.0
+  - @orderly.network/utils@1.5.1-alpha.0
+
 ## 1.5.0
 
 ### Minor Changes