+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @orderly.network/core
 
+## 1.5.1
+
+### Patch Changes
+
+- 34efc6c0: update
+- Updated dependencies [34efc6c0]
+  - @orderly.network/types@1.5.1
+
+## 1.5.1-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@1.5.1-alpha.0
+
 ## 1.5.0
 
 ### Minor Changes