+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/futures
 
+## 3.3.1
+
+### Patch Changes
+
+- 34efc6c0: update
+- Updated dependencies [34efc6c0]
+  - @orderly.network/types@1.5.1
+  - @orderly.network/utils@1.5.1
+
+## 3.3.1-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@1.5.1-alpha.0
+  - @orderly.network/utils@1.5.1-alpha.0
+
 ## 3.3.0
 
 ### Minor Changes